Change to micromath

+++ src/data_item.rs
@@ -31,11 +31,11 @@ use serde::{Deserialize, Serialize};
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone)]
 pub struct DataItem {
-    open: f64,
-    high: f64,
-    low: f64,
-    close: f64,
-    volume: f64,
+    open: f32,
+    high: f32,
+    low: f32,
+    close: f32,
+    volume: f32,
 }
 
 impl DataItem {
@@ -45,41 +45,41 @@ impl DataItem {
 }
 
 impl Open for DataItem {
-    fn open(&self) -> f64 {
+    fn open(&self) -> f32 {
         self.open
     }
 }
 
 impl High for DataItem {
-    fn high(&self) -> f64 {
+    fn high(&self) -> f32 {
         self.high
     }
 }
 
 impl Low for DataItem {
-    fn low(&self) -> f64 {
+    fn low(&self) -> f32 {
         self.low
     }
 }
 
 impl Close for DataItem {
-    fn close(&self) -> f64 {
+    fn close(&self) -> f32 {
         self.close
     }
 }
 
 impl Volume for DataItem {
-    fn volume(&self) -> f64 {
+    fn volume(&self) -> f32 {
         self.volume
     }
 }
 
 pub struct DataItemBuilder {
-    open: Option<f64>,
-    high: Option<f64>,
-    low: Option<f64>,
-    close: Option<f64>,
-    volume: Option<f64>,
+    open: Option<f32>,
+    high: Option<f32>,
+    low: Option<f32>,
+    close: Option<f32>,
+    volume: Option<f32>,
 }
 
 impl DataItemBuilder {
@@ -93,27 +93,27 @@ impl DataItemBuilder {
         }
     }
 
-    pub fn open(mut self, val: f64) -> Self {
+    pub fn open(mut self, val: f32) -> Self {
         self.open = Some(val);
         self
     }
 
-    pub fn high(mut self, val: f64) -> Self {
+    pub fn high(mut self, val: f32) -> Self {
         self.high = Some(val);
         self
     }
 
-    pub fn low(mut self, val: f64) -> Self {
+    pub fn low(mut self, val: f32) -> Self {
         self.low = Some(val);
         self
     }
 
-    pub fn close(mut self, val: f64) -> Self {
+    pub fn close(mut self, val: f32) -> Self {
         self.close = Some(val);
         self
     }
 
-    pub fn volume(mut self, val: f64) -> Self {
+    pub fn volume(mut self, val: f32) -> Self {
         self.volume = Some(val);
         self
     }
@@ -154,7 +154,7 @@ mod tests {
 
     #[test]
     fn test_builder() {
-        fn assert_valid((open, high, low, close, volume): (f64, f64, f64, f64, f64)) {
+        fn assert_valid((open, high, low, close, volume): (f32, f32, f32, f32, f32)) {
             let result = DataItem::builder()
                 .open(open)
                 .high(high)
@@ -165,7 +165,7 @@ mod tests {
             assert!(result.is_ok());
         }
 
-        fn assert_invalid((open, high, low, close, volume): (f64, f64, f64, f64, f64)) {
+        fn assert_invalid((open, high, low, close, volume): (f32, f32, f32, f32, f32)) {
             let result = DataItem::builder()
                 .open(open)
                 .high(high)
